Revert "Regex patterns"
