+++ Consumer.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PastriesDeliveryTypeSystem
+{
+    /// <summary>
+    /// Intended for working with a consumer.
+    /// </summary>
+    public class Consumer
+    {
+        public string Name { get; set; }
+        public string Surname { get; set; }
+        public string PhoneNumber { get; set; }
+        public string Adress { get; set; }
+    }
+    public enum PaymentMethod { Cash, Card, Smartphone }
+}
