+++ FoodDeliveryTask.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31129.286
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FoodDeliveryTask", "FoodDeliveryTask.csproj", "{94AFAC42-F5CC-4545-8D0C-37EB500EC345}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{94AFAC42-F5CC-4545-8D0C-37EB500EC345}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{94AFAC42-F5CC-4545-8D0C-37EB500EC345}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{94AFAC42-F5CC-4545-8D0C-37EB500EC345}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{94AFAC42-F5CC-4545-8D0C-37EB500EC345}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {976747DB-789A-46AA-8D8F-DC58715254BE}
+	EndGlobalSection
+EndGlobal
