--- PastriesDelivery/Constants/DiscountPercentEnum.cs
@@ -1,9 +1,0 @@
-﻿namespace PastriesDelivery
-{
-    public enum DiscountPercentEnum
-    {
-        TwentyUnits = 2,
-        FiftyUnits = 3,
-        HundredUnits = 4
-    }
-}
