--- PastriesDelivery/Constants/Role.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PastriesDelivery
-{
-    public enum Role
-    {
-        Provider,
-        Сustomer
-    }
-}
