--- PastriesDelivery/Contracts/ICustomerManager.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace PastriesDelivery
-{
-    public interface ICustomerManager
-    {
-        List<Product> ExtractProducts();
-    }
-}
