--- PastriesDelivery/Contracts/IDataDisplayer.cs
@@ -1,7 +1,0 @@
-﻿namespace PastriesDelivery
-{
-    public interface IDataDisplayer
-    {
-        void DisplayAvailableProducts();
-    }
-}
