--- PastriesDelivery/Contracts/IStorage.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace PastriesDelivery
-{
-    public interface IStorage
-    {
-        List<Product> Products { get; set; }
-        List<Order> Orders { get; set; }
-    }
-}
