--- PastriesDelivery/Managers/ConsumerManager.cs
@@ -1,14 +1,0 @@
-﻿using System.Linq;
-
-namespace PastriesDelivery
-{
-    /// <summary>
-    /// This class contains methods intended for work with consumer.
-    /// </summary>
-    public class ConsumerManager : СustomerManager, ICustomerManager
-    {
-        public ConsumerManager(IStorage storage) : base(storage)
-        {
-        }
-    }
-}
