+++ Provider.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PastriesDeliveryTypeSystem
+{
+    /// <summary>
+    /// Intended for work with providers of a delivery service.
+    /// </summary>
+    public class Provider
+    {
+        public string Name { get; set; }
+        public string Adress { get; set; }
+        public string PhoneNumber { get; set; }
+    }
+}
